LRQA-39649 Move methods and variables from BaseBatch to JunitBatch
